refactor: delete comments
