+++ src/ciphers/rotate-by-8-encrypt.js
@@ -18,7 +18,6 @@ const rotateBy8Encrypt = (text, mode) => {
         : char.charCodeAt(0),
     );
   const res = String.fromCharCode(...charCodeArr);
-  // console.log('charCodeArr: ', charCodeArr);
   console.log('before: ', text);
   console.log('result: ', res);
   return res;
