+++ src/helpers/get-flags.js
@@ -1,7 +1,3 @@
 const getFlags = () => process.argv.slice(2);
 
-// module.exports = {
-//   getFlags,
-// };
-
 export default getFlags;
