+++ src/streams/write-stream.js
@@ -1,18 +1,13 @@
 import fs from 'fs';
-import getOptionValue from '../get-options.js';
+import getOptionValue from '../helpers/get-options.js';
 
 const createWriteStream = () => {
   const outputValue = getOptionValue('-o');
   if (outputValue) {
     return fs.createWriteStream(outputValue);
   } else {
-    // console.log('Result: ');
     return process.stdout;
   }
 };
-
-// module.exports = {
-//   createWriteStream,
-// };
 
 export default createWriteStream;
